Several removals and cleanups:
Remove pyLaunchTime, which never quite worked because filesystem timestamps
could be wildly out of sync with the wall clock.

Remove reloadingService, which doesn't make any sense (detector is useful
as a process-level behavior; it has nothing to do with services).

Allow users to pass a logCallable instead of guessing whether to use
twisted.python.log or `print`.

Document some things a bit better.

A lot of code formatting cleanup (sorry).